samus: Enable update screen for software sync
Since the PD software sync is slow enable support for displaying
a screen telling the user that something is happening.

BUG=chrome-os-partner:32379
BRANCH=samus
TEST=manual testing:
1) in normal mode, with EC/PD in RW, ensure that they are rebooted
to RO and the VGA Option ROM is loaded and the wait screen is
displayed, and then the system is rebooted at the end and the
VGA Option ROM is not loaded.
2) same as #1 with EC/PD in RO already, same result
3) same as #1 with system in developer mode, same result except
there is no reboot at the end of software sync
4) same as #1 with system in developer mode and EC/PD in RO,
ensure that there is no extra reboot at the beginning or end of
software sync.

Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223850
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
(cherry picked from commit 6971d74ff50ced0ef94d5fec26c0e6a071d207b2)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: I125744f58c6b84df1af3943d9be98fe55c7117d5
Reviewed-on: https://chromium-review.googlesource.com/224362
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Duncan Laurie <dlaurie@chromium.org>