auron_paine: Delay the setup of pc-beep verbs
This is basicly clone from Leon
https://chromium-review.googlesource.com/#/c/180984/

auron_paine has an unwanted  audio pop sound on system resume.
The cause is that the ALC283 requires about 100ms between the
device reset and the setup of the pc-beep verbs. Specificly the
EAPD signal enable needs to be delayed. We don't want to wait
100ms in the azalia code, so add a callback to setup the pc-beep
verb after the coreboot table setup.

BUG=chrome-os-partner:31842
TEST=test the pop noise when s3 resume is gone
Signed-off-by: Pan Sheng-Liang <Sheng-Liang.Pan@quantatw.com>
Change-Id: Ia23c57c71a687ae850ed92595635f98b555d7986
Reviewed-on: https://chromium-review.googlesource.com/240116
Reviewed-by: Shawn N <shawnn@chromium.org>
Commit-Queue: Shawn N <shawnn@chromium.org>
Tested-by: Shawn N <shawnn@chromium.org>