soc/intel/common: Add function to protect MRC cache
Add support for applying write protection to the MRC cache
region in SPI flash.

This is only enabled if there is write protect GPIO that is
set, and the flash status register reports that the flash
chip is currently write protected.

Then it will call out to a SOC specific function that will
enable write protection on the RW_MRC_CACHE region of flash.

The implementation is not quite as clean as I would like because
there is not a common flash protect interface across SOCs so
instead it relies on a new Kconfig variable to be set that will
indicate a SOC implements the function to protect a region of
SPI flash.

BUG=chrome-os-partner:28234
BRANCH=broadwell
TEST=build and boot on samus
1) with either WPSW=0 or SRP0=0 the PRR is not applied
2) with both WPSW=1 and SRP0=1 the PRR is applied

Change-Id: I94e54e4723b1dcdacbb6a05f047d0c0ebc7d8711
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241170
Reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262075
Commit-Queue: Shawn N <shawnn@chromium.org>
Tested-by: Shawn N <shawnn@chromium.org>
(cherry picked from commit 93c656745062205c405ee88bb8a8cf4bbfca80a6)
Reviewed-on: https://chromium-review.googlesource.com/263390
Reviewed-by: Mike M Hsieh <mike.m.hsieh@intel.com>
Commit-Queue: Mike M Hsieh <mike.m.hsieh@intel.com>
Tested-by: Mike M Hsieh <mike.m.hsieh@intel.com>