Improve coreboot startup printouts
All coreboot phases apart from rambase add an empty line to the
console output at startup, which makes it easy to separate the stages
visually.

This patch adds an empty line before rambase section of the console
output starts.

While we are at it, also adding the absolute offset value to the CBFS
log, to make it easier to understand which particular CBFS section the
file is loaded from.

BRANCH=storm
BUG=none
TEST=rebooted a Whirlwind device, observed an empty line before the
     ramstage section of the log and absolute offsets reported by
     CBFS.

Change-Id: I5cc727127374d6e55b8ff6f45b250ef97125a8ec
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/255120
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261693