broadwell: Don't fail  on system_reset() in S3 resume
The system can't be reset until the Intel reference code runs, but
the flag is located in CBMEM and that region is not setup in a
S3 resume. In this case, do a soft_reset() to get back to a normal
boot flow.

BUG=chrome-os-partner:39297, chrome-os-partner:40560
BRANCH=auron
TEST=flashrom -p host -w image.bin
     Put the system into s3 and resume.
     System should reset and not hang.

Change-Id: Iaa1673a9c869bc502d6ce0584abb78782e3309cb
Signed-off-by: Marc Jones <marc.jones@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/273633
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
(cherry picked from commit 9d3a1ca7438036547972be020a7064819fb9ced9)
Reviewed-on: https://chromium-review.googlesource.com/273670
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>