strago: fix touchscreen gpio
Atmel touchscreen driver expects the first gpio resource in asl
for the reset line. But Strago does not use reset gpio line.
Hence driver will assume that the gpio based irq line to be the reset line
and changes the direction to output.
This will cause irq registration to fail.
Solution is to pass Interrupt resource instread of GpioInt

BRANCH=none
BUG=chrome-os-partner:41791
TEST=verify touchscreen functionality

Change-Id: I02f207ca42c5d4b93270e04565474f05d7d4135a
Signed-off-by: Jagadish Krishnamoorthy <jagadish.krishnamoorthy@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/283575
Reviewed-by: Aaron Durbin <adurbin@chromium.org>